+++ src/constellation/DNS.go
@@ -179,11 +179,11 @@ func InitDNS() {
 	if(!config.ConstellationConfig.DNSDisabled) {
 		go (func() {
 			dns.HandleFunc(".", handleDNSRequest)
-			server := &dns.Server{Addr: ":" + DNSPort, Net: "udp"}
+			server := &dns.Server{Addr: "192.168.201.1:" + DNSPort, Net: "udp"}
 
 			utils.Log("Starting DNS server on :" + DNSPort)
 			if err := server.ListenAndServe(); err != nil {
-				utils.Fatal("Failed to start server: %s\n", err)
+				utils.Error("Failed to start DNS server", err)
 			}
 		})()
 	}