+++ src/user/register.go
@@ -50,7 +50,8 @@ func UserRegister(w http.ResponseWriter, req *http.Request) {
 			return
 		}
 
-		c, errCo := utils.GetCollection(utils.GetRootAppId(), "users")
+		c, closeDb, errCo := utils.GetEmbeddedCollection(utils.GetRootAppId(), "users")
+  defer closeDb()
 		if errCo != nil {
 				utils.Error("Database Connect", errCo)
 				utils.HTTPError(w, "Database", http.StatusInternalServerError, "DB001")